@@ -1,15 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFrameworks>netcoreapp3.1;net5.0;net6.0;net7.0</TargetFrameworks>
+        <TargetFrameworks>net5.0;net6.0;net7.0</TargetFrameworks>
         <IsPackable>false</IsPackable>
         <ApplicationIcon />
         <OutputType>Exe</OutputType>
-        <RootNamespace>Tests</RootNamespace>
-        <StartupObject>Tests.Program</StartupObject>
+        <RootNamespace>Test</RootNamespace>
         <AssemblyName>Test</AssemblyName>
-        <NoWarn>IDE0130;CS1591;CA1806;CA1825;</NoWarn>
-        <LangVersion>preview</LangVersion>
+        <WarningLevel>0</WarningLevel>
+        <StartupObject>Test.Program</StartupObject>
     </PropertyGroup>
 
     <ItemGroup>
@@ -17,10 +16,10 @@
         <PackageReference Include="Microsoft.TestPlatform.TestHost" Version="17.6.1" />
         <PackageReference Include="NUnit" Version="3.13.3" />
         <PackageReference Include="NUnit3TestAdapter" Version="4.5.0" />
-        <PackageReference Include="System.Text.Encoding.CodePages" Version="7.0.0" />
     </ItemGroup>
 
     <ItemGroup>
         <ProjectReference Include="..\Cave.IO\Cave.IO.csproj" />
     </ItemGroup>
+
 </Project>
