@@ -1,90 +1,124 @@
-﻿using System;
+﻿#if NETCOREAPP1_0_OR_GREATER && !NETCOREAPP2_0_OR_GREATER
+#define ALTERNATE_CODE
+#endif
+
+using System;
 using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using System.Reflection;
+using System.Runtime.Versioning;
 using NUnit.Framework;
 
-namespace Tests
-{
-    class Program
-    {
-        #region Private Methods
+namespace Test;
 
-        static int Main(string[] args)
+[ExcludeFromCodeCoverage]
+class Program
+{
+    public const bool Verbose = false;
+
+    static int Main(string[] args)
+    {
+        var errors = 0;
+
+#if ALTERNATE_CODE
+        var asm = typeof(Program).GetTypeInfo().Assembly;
+        var targetFramework = asm.GetCustomAttributes<TargetFrameworkAttribute>().FirstOrDefault();
+        var frameworkVersion = targetFramework.FrameworkDisplayName;
+        var types = asm.DefinedTypes.Select(t => t.AsType());
+#else
+        var types = typeof(Program).Assembly.GetTypes().ToArray();
+        var frameworkVersion = "net " + Environment.Version;
+#endif
+        Console.ForegroundColor = ConsoleColor.Yellow;
+        Console.WriteLine(frameworkVersion);
+        Console.WriteLine();
+        foreach (var type in types)
         {
-            Verbose = args.Contains("-v") || args.Contains("--verbose");
-            var errors = 0;
-            var types = typeof(Program).Assembly.GetTypes();
-            foreach (var type in types.OrderBy(t => t.Name))
+            Console.ResetColor();
+#if ALTERNATE_CODE
+            var attrib = type.GetTypeInfo().GetCustomAttribute<TestFixtureAttribute>();
+            if (attrib is not TestFixtureAttribute) continue;
+#else
+            var typeAttributes = type.GetCustomAttributes(typeof(TestFixtureAttribute), false).ToArray();
+            var typeAttributesCount = typeAttributes.Length;
+            if (typeAttributesCount == 0)
             {
-                if (!type.GetCustomAttributes(typeof(TestFixtureAttribute), false).Any())
+                continue;
+            }
+#endif
+
+            Console.WriteLine("Create " + type);
+            var instance = Activator.CreateInstance(type);
+            var methods = type.GetMethods();
+            foreach (var method in methods)
+            {
+                var methodAttributes = method.GetCustomAttributes(typeof(TestAttribute), false).ToArray();
+
+#if ALTERNATE_CODE
+                var methodAttributesCount = methodAttributes.Count();
+#else
+                var methodAttributesCount = methodAttributes.Length;
+#endif
+                if (methodAttributesCount == 0)
                 {
                     continue;
                 }
 
-                var instance = Activator.CreateInstance(type);
-                foreach (var method in type.GetMethods())
+                GC.Collect(999, GCCollectionMode.Forced);
+
+                Console.ForegroundColor = ConsoleColor.Cyan;
+                Console.WriteLine($"{method.DeclaringType.Name}.cs: info TI0001: Start {method.Name}: {frameworkVersion}");
+                Console.ResetColor();
+                try
                 {
-                    if (!method.GetCustomAttributes(typeof(TestAttribute), false).Any())
-                    {
-                        continue;
-                    }
-
-                    GC.Collect(999, GCCollectionMode.Default);
-                    Console.ForegroundColor = ConsoleColor.Cyan;
-                    Console.WriteLine($"{method.DeclaringType.Name}.cs: info TI0001: Start {method.Name} framework {Environment.Version}");
+                    method.Invoke(instance, null);
+                    Console.ForegroundColor = ConsoleColor.Green;
+                    Console.WriteLine($"{method.DeclaringType.Name}.cs: info TI0002: Success {method.Name}: {frameworkVersion}");
                     Console.ResetColor();
-                    try
-                    {
-                        method.Invoke(instance, null);
-                        Console.ForegroundColor = ConsoleColor.Green;
-                        Console.WriteLine($"{method.DeclaringType.Name}.cs: info TI0002: Success {method.Name} framework {Environment.Version}");
-                        Console.ResetColor();
-                    }
-                    catch (Exception ex)
-                    {
-                        Debug.WriteLine(ex);
-                        Console.ForegroundColor = ConsoleColor.Red;
-                        Console.WriteLine($"{method.DeclaringType.Name}.cs: error TE0001: {ex.Message} framework {Environment.Version}");
-                        Console.WriteLine(ex);
-                        Console.ResetColor();
-                        errors++;
-                    }
-
-                    Console.WriteLine("---");
                 }
+                catch (Exception ex)
+                {
+                    Debug.WriteLine(ex);
+                    Console.ForegroundColor = ConsoleColor.Red;
+                    Console.WriteLine($"{method.DeclaringType.Name}.cs: error TE0001: {ex.Message}: {frameworkVersion}");
+                    Console.WriteLine(ex.ToString());
+                    Console.WriteLine(ex.StackTrace);
+                    Console.ResetColor();
+                    errors++;
+                }
+                Console.WriteLine("---");
             }
-
-            if (errors == 0)
-            {
-                Console.ForegroundColor = ConsoleColor.Green;
-                Console.WriteLine($"---: info TI9999: All tests successfully completed at framework {Environment.Version}.");
-            }
-            else
-            {
-                Console.ForegroundColor = ConsoleColor.Red;
-                Console.WriteLine($"---: error TE9999: {errors} tests failed framework {Environment.Version}!");
-            }
-
-            Console.ResetColor();
-            if (Debugger.IsAttached)
-            {
-                WaitExit();
-            }
-
-            return errors;
         }
-
-        static void WaitExit()
+        if (errors == 0)
         {
-            Console.Write("--- press enter to exit ---");
-            while (Console.ReadKey(true).Key != ConsoleKey.Enter)
-            {
-                ;
-            }
+            Console.ForegroundColor = ConsoleColor.Green;
+            Console.WriteLine($"---: info TI9999: All tests successfully completed: {frameworkVersion}");
+        }
+        else
+        {
+            Console.ForegroundColor = ConsoleColor.Red;
+            Console.WriteLine($"---: error TE9999: {errors} tests failed: {frameworkVersion}");
+        }
+        Console.ResetColor();
+        if (Debugger.IsAttached)
+        {
+            WaitExit();
         }
 
-        public static bool Verbose { get; private set; }
-
-        #endregion Private Methods
+        return errors;
     }
+
+#if ALTERNATE_CODE
+    static void WaitExit() { }
+#else
+    static void WaitExit()
+    {
+        Console.Write("--- press enter to exit ---");
+        while (Console.ReadKey(true).Key != ConsoleKey.Enter)
+        {
+            ;
+        }
+    }
+#endif
 }
